DFT: Refactor code structure
Now we compile DFT as a set of libraries
