+++ src/dft/src/architect/ScanArchitectHeuristic.hh
@@ -31,7 +31,7 @@
 // POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
-#include "architect/ScanArchitect.hh"
+#include "ScanArchitect.hh"
 
 namespace dft {
