+++ src/dft/src/config/CMakeLists.txt
@@ -1,0 +1,19 @@
+add_library(dft_config_lib
+  # Keep sorted
+  DftConfig.cpp
+  ScanArchitectConfig.cpp
+)
+
+target_include_directories(dft_config_lib
+  PUBLIC
+    ${CMAKE_CURRENT_LIST_DIR}
+)
+
+target_link_libraries(dft_config_lib
+  PRIVATE
+    odb
+    dbSta
+    OpenSTA
+    utl
+    dft_utils_lib
+)
