+++ src/dft/test/cpp/CMakeLists.txt
@@ -5,8 +5,13 @@ include("openroad")
 set(DFT_HOME ${PROJECT_SOURCE_DIR})
 
 set(TEST_LIBS
+  utl
   dbSta
-  dft)
+  dft_cells_lib
+  dft_config_lib
+  dft_architect_lib
+  dft_clock_domain_lib
+)
 
 add_executable(TestScanArchitectHeuristic TestScanArchitectHeuristic.cpp ScanCellMock.cpp)
 target_link_libraries(TestScanArchitectHeuristic ${TEST_LIBS})
